Move manga list component to containers directory
