+++ src/styles/Error.module.css
@@ -3,7 +3,6 @@
   justify-content: center;
   align-items: center;
   text-align: center;
-  flex: 1;
 }
 
 .error img {
